Make read, write, and open be 'free' functions.  Justification...
These are not intrinsic to every kind of file descriptor.  If open is a
member, so should epoll_create, socket, and signalfd (among others).  But
that would require dragging in a whole bunch of constants and types
associated solely with each of these calls. That doesn't seem right, the
fd type should be thin.

So, the fd type will now contain only a very few functions that are relevant
for just about any kind of file descriptor.

Of course, the magic of argument dependent lookup enables some of these now
'free' functions (like read and write) to be called without referencing the
namespace they're in because C++ rightly considers them part of the
interface for the fd type they take.